YUK-6284, New customer account and device enrollment import process
-	Fixed a conversion issue with programEnrollment using a double instead of a float.
-  Fixed programSignUpAction to also update if the relay number has been changed on an existing enrollment.

+++ yukon-client/stars/src/com/cannontech/stars/dr/enrollment/dao/impl/EnrollmentDaoImpl.java
@@ -70,7 +70,7 @@ public class EnrollmentDaoImpl implements EnrollmentDao {
                 programEnrollment.setApplianceCategoryId(rs.getInt("applianceCategoryId"));
                 programEnrollment.setProgramId(rs.getInt("programId"));
                 programEnrollment.setLmGroupId(rs.getInt("lmGroupId"));
-                programEnrollment.setApplianceKW(rs.getDouble("KWCapacity"));
+                programEnrollment.setApplianceKW(rs.getFloat("KWCapacity"));
                 programEnrollment.setRelay(rs.getInt("relay"));
                 programEnrollment.setInventoryId(rs.getInt("inventoryId"));
                 return programEnrollment;
